feat(restaurant): add entity and profile endpoints

+++ gastroflow/src/config/typeorm.ts
@@ -1,19 +1,21 @@
 import { registerAs } from '@nestjs/config';
-import { environment } from './enviroment';
 
-export const typeOrmConfig = registerAs('typeorm', () => ({
-  type: 'postgres',
-  url: environment.DATABASE_URL,
-  ssl: {
-    rejectUnauthorized: false,
-  },
-  extra: {
-    ssl: {
-      rejectUnauthorized: false,
-    },
-  },
-  autoLoadEntities: true,
-  synchronize: true,
-  logging: true,
-  dropSchema: false,
-}));
+export const typeOrmConfig = registerAs('typeorm', () => {
+  const databaseUrl = process.env.DATABASE_URL;
+
+  if (!databaseUrl) {
+    throw new Error(
+      'DATABASE_URL no está definida en las variables de entorno',
+    );
+  }
+
+  return {
+    type: 'postgres' as const,
+    url: databaseUrl,
+    ssl: false,
+    autoLoadEntities: true,
+    synchronize: true,
+    logging: true,
+    dropSchema: false,
+  };
+});
